pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "A local AI voice assistant with real-time conversation capabilities"
requires-python = ">=3.10"
dependencies = [
    # Core dependencies
    "numpy==1.26.4",
    "scipy==1.14.0",
    "Pillow==10.4.0",
    "soundfile==0.12.1",
    
    # Audio I/O
    "PyAudio==0.2.14",
    "opuslib==3.0.1",
    
    # ML/AI frameworks (torch must be before flash-attn)
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "onnxruntime==1.20.1",
    "transformers==4.46.2",
    "accelerate==1.3.0",
    "deepspeed==0.16.3",
    "huggingface-hub==0.28.1",
    
    # Performance optimization (requires torch to be installed first)
    # "flash-attn==2.7.4.post1",  # Skipped - requires CUDA Toolkit to build
    
    # LLM
    "llama-cpp-python==0.3.4",
    "ninja==1.11.1.3",
    
    # TTS
    "kokoro==0.7.3",
    "coqui-tts==0.25.3",
]

[tool.uv]
# Development dependencies (optional)
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }
torchvision = { index = "pytorch" }

# Override flash-attn to disable build isolation
[tool.uv.pip]
no-build-isolation-package = ["flash-attn"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Tell hatchling this is an application, not a library
[tool.hatch.build.targets.wheel]
packages = ["."]